feat: implement k_nearest_interpolation; refactor;
