+++ samples/experimental/2-template-shifts/alignment/config.json
@@ -6,6 +6,6 @@
     ],
     "save_detections": true,
     "save_image_level": 0,
-    "show_image_level": 5
+    "show_image_level": 3
   }
 }
